Adding Pid to new_data message
Not having the PO Box's pid in the new_data message means that the
current PO Box implementation can only be used safely with a single box
per process in notify mode. Active mode already contains the pid, and so
is safe to use with multiple PO Boxes.

This commit adds the pid of the po box to the new_data message so that a
process owning multiple PO Boxes can know which one sent the mail
notification so it can safely be polled.