Fix accounting error resulting in queue crash
When manually dropping a message (with the activation filter), the
dropped count is not subtracted from the buffer size. This leads to
invalid actions being taken, and in the case of queue buffers, crashes.

This patch makes it so the manually dropped messages are accounted for
and should fix the bug.